@@ -4,20 +4,20 @@
  */
 
 import React, { useEffect, useState } from 'react';
-import { connect } from 'react-redux';
+import { useSelector, useDispatch } from 'react-redux';
 
-import type { State } from '../reducers';
 import useWindowSize from '../utils/reactHookResize';
 import { showChatModal } from '../actions';
 
 import Chat from './Chat';
 
 
-function ChatBox({
-  chatOpen,
-  triggerModal,
-}) {
+const ChatBox = () => {
   const [render, setRender] = useState(false);
+
+  const chatOpen = useSelector((state) => state.modal.chatOpen);
+
+  const dispatch = useDispatch();
 
   useEffect(() => {
     window.setTimeout(() => {
@@ -31,7 +31,7 @@ function ChatBox({
 
   const { width } = useWindowSize();
   if (width < 604 && chatOpen) {
-    triggerModal();
+    dispatch(showChatModal(true));
   }
 
   return (
@@ -44,19 +44,6 @@ function ChatBox({
       </div>
     )
   );
-}
+};
 
-function mapStateToProps(state: State) {
-  const { chatOpen } = state.modal;
-  return { chatOpen };
-}
-
-function mapDispatchToProps(dispatch) {
-  return {
-    triggerModal() {
-      dispatch(showChatModal(true));
-    },
-  };
-}
-
-export default connect(mapStateToProps, mapDispatchToProps)(ChatBox);
+export default React.memo(ChatBox);
